requirements.txt
netcdf4
xarray
metpy
herbie-data
dask
dask-jobqueue